Refactor article tab component logic into its own hook
