+++ packages/help-center/src/components/help-center-article.tsx
@@ -1,9 +1,10 @@
 import { recordTracksEvent } from '@automattic/calypso-analytics';
-import { useEffect, createInterpolateElement, useCallback, useState } from '@wordpress/element';
+import { useEffect, createInterpolateElement } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 import { useHelpCenterContext } from '../contexts/HelpCenterContext';
 import { usePostByUrl } from '../hooks';
+import { useHelpCenterArticleTabComponent } from '../hooks/use-help-center-article-tab-component';
 import { BackToTopButton } from './back-to-top-button';
 import ArticleContent from './help-center-article-content';
 
@@ -12,13 +13,11 @@ import './help-center-article.scss';
 export const HelpCenterArticle = () => {
 	const [ searchParams ] = useSearchParams();
 	const { sectionName } = useHelpCenterContext();
-	const location = useLocation();
-	const navigate = useNavigate();
-	const [ tabHash, setTabHash ] = useState( '' );
 	const postUrl = searchParams.get( 'link' ) || '';
 	const query = searchParams.get( 'query' );
 
 	const { data: post, isLoading, error } = usePostByUrl( postUrl );
+	useHelpCenterArticleTabComponent( post?.content );
 
 	useEffect( () => {
 		//If a url includes an anchor, let's scroll this into view!
@@ -54,66 +53,6 @@ export const HelpCenterArticle = () => {
 				: recordTracksEvent( 'calypso_helpcenter_article_viewed', tracksData );
 		}
 	}, [ post, query, sectionName ] );
-
-	const toggleTab = ( element: Element, show: boolean ) => {
-		( element as HTMLElement ).style.display = show ? 'block' : 'none';
-		element.setAttribute( 'aria-hidden', show ? 'false' : 'true' );
-	};
-
-	const toggleTabTitle = ( element: Element, show: boolean ) => {
-		element.setAttribute( 'aria-selected', show ? 'true' : 'false' );
-	};
-
-	const activateTab = useCallback( () => {
-		const hash = tabHash;
-
-		const tabs = Array.from( document.querySelectorAll( '.wp-block-wpsupport3-tabs' ) );
-
-		tabs.forEach( ( tab ) => {
-			const titles = Array.from( tab.querySelectorAll( '.wpsupport3-tab__title' ) );
-			const bodies = Array.from(
-				tab.querySelectorAll( '.wp-block-wpsupport3-tab:not(.invisible_tabpanel)' )
-			);
-
-			const match = titles.findIndex( ( titles ) => titles.id === hash?.substring( 1 ) );
-
-			// Reset selection
-			titles.forEach( ( title ) => toggleTabTitle( title, false ) );
-			bodies.forEach( ( body ) => toggleTab( body, false ) );
-
-			if ( hash && match !== -1 ) {
-				toggleTabTitle( titles[ match ], true );
-				toggleTab( bodies[ match ], true );
-			} else {
-				// If the first tab is invisible from the editor, we set the first tab as active.
-				toggleTabTitle( titles[ 0 ], true );
-				toggleTab( bodies[ 0 ], true );
-			}
-		} );
-	}, [ tabHash ] );
-
-	useEffect( () => {
-		if ( tabHash || post ) {
-			activateTab();
-		}
-	}, [ activateTab, tabHash, post ] );
-
-	useEffect( () => {
-		if ( post ) {
-			const titles = Array.from(
-				document.querySelectorAll( '.wp-block-wpsupport3-tabs .wpsupport3-tab__title' )
-			);
-			titles.forEach( ( title ) => {
-				title.addEventListener( 'click', ( e ) => {
-					e.preventDefault();
-					setTabHash( `#${ title?.id }` );
-					setTimeout( () => {
-						window.scroll( 0, document.documentElement.scrollTop );
-					} );
-				} );
-			} );
-		}
-	}, [ location.pathname, location.search, navigate, post ] );
 
 	return (
 		<div className="help-center-article">
